+++ src/py_gql/utilities/__init__.py
@@ -20,7 +20,6 @@ from .collect_fields import (
     selected_fields,
 )
 from .introspection_query import introspection_query
-from .max_depth import MaxDepthValidationRule
 from .untyped_value_from_ast import untyped_value_from_ast
 from .value_from_ast import value_from_ast
 
@@ -38,5 +37,4 @@ __all__ = (
     "introspection_query",
     "untyped_value_from_ast",
     "value_from_ast",
-    "MaxDepthValidationRule",
 )